matrix: move to math/
