+++ meson.build
@@ -131,7 +131,6 @@ sources = [
   'src/game/lara/lara_state.c',
   'src/game/level.c',
   'src/game/lot.c',
-  'src/game/matrix.c',
   'src/game/music.c',
   'src/game/objects/ai/ape.c',
   'src/game/objects/ai/bacon_lara.c',
@@ -243,6 +242,7 @@ sources = [
   'src/json/json_parse.c',
   'src/json/json_write.c',
   'src/log.c',
+  'src/math/matrix.c',
   'src/memory.c',
   'src/specific/s_audio.c',
   'src/specific/s_audio_sample.c',
