+++ src/game/output.c
@@ -4,10 +4,10 @@
 #include "3dsystem/phd_math.h"
 #include "config.h"
 #include "game/clock.h"
-#include "game/matrix.h"
 #include "game/random.h"
 #include "game/viewport.h"
 #include "global/vars.h"
+#include "math/matrix.h"
 #include "specific/s_misc.h"
 #include "specific/s_output.h"
 #include "specific/s_shell.h"
